requirements.txt
Python 3.10.6
Flask 2.2.0
Werkzeug 2.2.1
bootstrap 5.3.0
Flask-MySQL 1.5.2
flask-mysqldb 1.0.1
pyyaml 6.0.1